pyproject.toml
[project]
name = "enhanced-discord-bot-llms"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "outcastgeek", email = "outcastgeek+gitkraken@gmail.com" }
]
dependencies = [
    "openai>=1.28.0",
    "discord-py>=2.3.2",
    "instructor>=1.2.6",
    "groq>=0.6.0",
    "granian[all]>=1.3.1",
    "litestar[brotli,cryptography,minijinja,standard]>=2.8.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.4",
    "isort>=5.13.2",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "jupyterlab>=4.2.0",
    "httpie>=3.2.2",
]

[tool.rye.scripts]
gaou_dev = { cmd = "litestar --app 'enhanced_discord_bot_llms.gaouapp:app' run -r -d" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/enhanced_discord_bot_llms"]
